server/pyproject.toml
[tool.poetry]
name = "djangopress"
version = "0.0.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [{ include = "djangopress" }]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
psycopg = "^3.1.16"
dj-database-url = "^2.1.0"
gunicorn = "^21.2.0"
sentry-sdk = {extras = ["django"], version = "^1.39.1"}
isort = "^5.13.2"
djrender = "^0.1b7"
whitenoise = "^6.6.0"
pillow = "^10.3.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"

[tool.isort]
profile = "black"
known_first_party = ["djangopress"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
